requirements.txt
python-docx
xlwings
docxtpl
openpyxl
py2exe
aspose-words
os
datetime